reader_app/pyproject.toml
[project]
name = "reader3"
version = "0.1.0"
description = "Simple EPUB reader web app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ebooklib>=0.20",
    "fastapi>=0.121.2",
    "google-generativeai>=0.3.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]
